transitioning to C++ implementation
