+++ .gitignore
@@ -1,6 +1,6 @@
 build/
+src/CMakeUserPresets.json
 defaults.cfg
 main.plist
 src/test_data/
 src/sample_data/
-src/CMakeUserPresets.json
