+++ CMakeLists.txt
@@ -9,17 +9,14 @@ cmake_policy(SET CMP0091 NEW)
 project(
 	astrosight
 	VERSION 0.1.0
-	LANGUAGES C CXX
+	LANGUAGES CXX
 	DESCRIPTION "Image processing application for astrophotography.")
 
-set(CMAKE_C_STANDARD 17)
-set(CMAKE_C_STANDARD_REQUIRED TRUE)
-set(CMAKE_C_EXTENSIONS OFF)
-set(CMAKE_C_VISIBILITY_PRESET hidden)
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED TRUE)
 set(CMAKE_CXX_EXTENSIONS OFF)
 set(CMAKE_CXX_VISIBILITY_PRESET hidden)
+
 set(CMAKE_EXPORT_COMPILE_COMMANDS TRUE)
 set(CMAKE_SKIP_BUILD_RPATH FALSE)
 set(CMAKE_BUILD_WITH_INSTALL_RPATH FALSE)
@@ -33,6 +30,9 @@ option(BUILD_TESTING "Build tests" FALSE)
 option(BUILD_SHARED_LIBS "Build shared libraries" FALSE)
 option(BUILD_WITH_MT "Build libraries as MultiThreaded DLL (Windows Only)" FALSE)
 
+add_compile_options(
+  "-O3" "-Wall" "-Wpedantic" "-Wextra"
+)
 
 include(CTest)
 include(GNUInstallDirs)
