+++ README.md
@@ -2,15 +2,25 @@
 
 Image processing library for astrophotography
 
----------------------------------------------
+---
 
 [![Latest Release][release-badge]][[release-url]]
 [![License][license-badge]](LICENSE)
 [![CI Status][ci-badge]][ci-url]
 
 [release-badge]: https://img.shields.io/github/v/release/austinlucaslake/astrosight
-[release-url]: https://github.com/austinlucaslake/provision/releases/latest
+[release-url]: https://github.com/austinlucaslake/astrosight/releases/latest
 [license-badge]: https://img.shields.io/github/license/austinlucaslake/astrosight
-[ci-badge]: https://github.com/austinlucaslake/astrosight/actions/workflows/ci.yml/badge.svg
+[ci-badge]: https://github.com/austinlucaslake/astrosight/actions/workflows/ci.yaml/badge.svg
 [ci-url]: https://github.com/austinlucaslake/astrosight/actions
+
+---
+
+## Build
+
+Assuming `Conan` is installed, you may build this project as follows:
+`conan build . -b=missing`
+
+It can then be run using by executing the following script from the project's root directory:
+`./build/Release/main`
