--- src/test.cpp
@@ -1,3 +1,0 @@
-int main(void) {
-  return 1;
-}
